-removed uint -removed most of Poco (will hopefully remove the dependency) -removed a quite retarded Timer that tried to simulate vsync making everything slow -replaced all of the syncrhonization queues with the new lockless Pipe

+++ include/dojo/Array.h
@@ -91,7 +91,7 @@ namespace Dojo
 		 */
 		Array(int firstPageSize = 64, int newPageSize = 64, int validElements = 0)
 		{
-			DEBUG_ASSERT( validElements <= (size_t)firstPageSize, "The first page must contain at least 'validElements' elements" );
+			DEBUG_ASSERT( validElements <= firstPageSize, "The first page must contain at least 'validElements' elements" );
 			
 			pageSize = newPageSize;			
 			elements = validElements;
